requirements.txt
feedparser>=6.0
python-dotenv>=1.0
python-telegram-bot>=21.0
pytest>=8.0
ruff>=0.4
